Call the correct overload when not passing nominalTS
With the current code when nominalTS is not passed explicitly, instead
of calling the new C++ constructor, the wrapper calls the old C++
constructor and passes a null YieldTermStructureHandle, which makes
bootstrapping fail. This is how default arguments in SWIG work when
%kwargs feature is enabled. When it's not enabled, default arguments are
translated into overloads, so the correct code would have been
generated, but this seems quite subtle, so adding an explicit overload
seems better.